+++ src/Exception/File/EmptyFilePathException.php
@@ -25,6 +25,6 @@ class EmptyFilePathException extends Exception
      */
     public static function create(): EmptyFilePathException
     {
-        return new static('Path of the file is empty. Did you provide path of proper file?');
+        return new self('Path of the file is empty. Did you provide path of proper file?');
     }
 }